streamlit_app/requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
altair==5.1.2
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
seaborn==0.13.0
matplotlib==3.8.2
openai==1.3.7
langchain==0.0.350
chromadb==0.4.18
psycopg2-binary==2.9.9
redis==5.0.1
jwt==1.3.1
PyJWT==2.8.0
bcrypt==4.1.2
